requirements.txt
aiogram==3.16.0
psycopg2==2.9.10
rapidfuzz==3.11.0
PyYAML==6.0.2